+++ packages/shared/src/types/organizationMembership.ts
@@ -23,7 +23,7 @@ declare global {
   /**
    * If you want to provide custom types for the organizationMembership.publicMetadata
    * object, simply redeclare this rule in the global namespace.
-   * Every organizationMembership object will use the provided type.
+   * Every OrganizationMembership object will use the provided type.
    */
   interface OrganizationMembershipPublicMetadata {
     [k: string]: unknown;
@@ -32,7 +32,7 @@ declare global {
   /**
    * If you want to provide custom types for the organizationMembership.publicMetadata
    * object, simply redeclare this rule in the global namespace.
-   * Every organizationMembership object will use the provided type.
+   * Every OrganizationMembership object will use the provided type.
    */
   interface OrganizationMembershipPrivateMetadata {
     [k: string]: unknown;
@@ -40,7 +40,7 @@ declare global {
 }
 
 /**
- * The `OrganizationMembership` object is the model around an organization membership entity and describes the relationship between users and organizations.
+ * The `OrganizationMembership` object is the model around an Organization membership entity and describes the relationship between users and Organizations.
  * @interface
  */
 export interface OrganizationMembershipResource extends ClerkResource {
@@ -68,9 +68,9 @@ export type OrganizationCustomPermissionKey = ClerkAuthorization extends Placeho
   : Base['permission'];
 
 /**
- * `OrganizationCustomRoleKey` is a type that represents the user's role in an organization. It will be string unless the developer has provided their own types through [`ClerkAuthorization`](https://clerk.com/docs/guides/development/override-clerk-types-interfaces#example-custom-roles-and-permissions).
+ * `OrganizationCustomRoleKey` is a type that represents the user's Role in an Organization. It will be string unless the developer has provided their own types through [`ClerkAuthorization`](https://clerk.com/docs/guides/development/override-clerk-types-interfaces#example-custom-roles-and-permissions).
  *
- * Clerk provides the [default roles](https://clerk.com/docs/guides/organizations/roles-and-permissions#default-roles) `org:admin` and `org:member`. However, you can create [custom roles](https://clerk.com/docs/guides/organizations/roles-and-permissions#custom-roles) as well.
+ * Clerk provides the [default Roles](https://clerk.com/docs/guides/organizations/control-access/roles-and-permissions#default-roles) `org:admin` and `org:member`. However, you can create [Custom Roles](https://clerk.com/docs/guides/organizations/control-access/roles-and-permissions#custom-roles) as well.
  *
  * @interface
  */
@@ -90,8 +90,8 @@ export type OrganizationSystemPermissionKey =
   | `${OrganizationSystemPermissionPrefix}domains:read`;
 
 /**
- * OrganizationPermissionKey is a combination of system and custom permissions.
- * System permissions are only accessible from FAPI and client-side operations/utils
+ * OrganizationPermissionKey is a combination of System and Custom Permissions.
+ * System Permissions are only accessible from FAPI and client-side operations/utils
  */
 export type OrganizationPermissionKey = ClerkAuthorization extends Placeholder
   ? ClerkAuthorization['permission'] extends string